refactor: move PostHog instance to diagnostics module and update documentation
